(refactor): move api generators
