--- src/generators/src/infra/server.js
@@ -1,7 +1,0 @@
-module.exports = async ({ template: { generate }, parameters: { options } }) => async () => {
-  await generate({
-    template: 'infra/api/server.ejs',
-    target: 'src/infra/api/server.js',
-    props: { ...options }
-  })
-}
